페이징처리 및 관리자기능 추가중

+++ weatherAPIpractice/src/main/java/co/mg/vo/MemberVO.java
@@ -11,6 +11,6 @@ public class MemberVO {
 	private String nickname;
 	private String pw;
 	private Date signupDate;
-	private int isActive;
+	private byte isActive;
 	private String member_grade;
 }
